Add support for default values
This update provides the `_color()` mixin to support functionality that's similar to Sass' `!default` flag.

You can do this by passing `default` as the final argument when you define a color:

```scss
@include _color((
   orange: (
      default: orange,
      red: orangered,
   ),
), default);
```

The default color scheme has been updated so that the colors are set with this `default` setting.
This ensures the `seed-color-scheme` pack can be `@import` without fear of unexpected overrides.

This change is important as it allows other packs to create flexible palettes that allow users to easily override default colors.