Remove Barista for testing (for now)
This update removes the Barista testing phase for the Seed packs.
This improves the workflow for folks to work on Seed.

Would like to introduce testing again, but to do it at a global level, instead
of per package. Would also like to use Jest, instead of mocha for testing.